Add 'FoodStoreRequest' class for validation rules for store method.

+++ database/migrations/2019_11_30_104013_create_foods_table.php
@@ -19,7 +19,7 @@ class CreateFoodsTable extends Migration
             $table->string('brand')->nullable();
             $table->integer('serving')->default(0);
             $table->enum('unit', ['gram', 'milliliter', 'kilogram', 'liter'])->default('gram');
-            $table->decimal('calories', 5, 1)->default(0);
+            $table->integer('calories')->default(0);
             $table->decimal('fat', 5, 1)->default(0);
             $table->decimal('fat_satured', 5, 1)->default(0);
             $table->decimal('cholesterol', 5, 1)->default(0);
